seperate queries and run formatters
